Load remote images in source-umbraco plugin using gatsby-source-filesystem
This commit is mostly to make sure my progress for today is in source control.

Definetly needs refactoring
Also needs to make imageKeys configurable
Also might see if it is possible to remove the dependency on gatsby-source-filesystem` within our actual gatsby site. So we only need it in our custom local plugin.