Load arrays of remote files too in 'load-remote-files'
The code resembles that of 'handle-foreign-key-fields' since both features do something similar. (Scan for certain fields, replace those fields with something else. The "something else" is just different.)

Before, the plugin would warn you if a remote file could not be loaded and then move on, keeping the old property and value. I've updated this to just crash with some debug information. Since it get's weird when you fail to load 1 file in an array, I think that when that happens it's most likely a mistake somewhere.

I can't think of any use cases where a ___FILE field would/should contain something other than a link to a file.